requirements.txt
aiohttp 
rgbprint
requests
pillow